test(editor): fix tests failing
While upgrading @apollo/client so that all libraries/dependencies use the same version the tests started to fail.
The errors were that they tried to write a field but it did not exist in the result. This seems to be related to this: https://github.com/apollographql/apollo-client/issues/8677. I don't know if this is a bug by apollo, but it seems like it.

Another error encountered was that `value` that was given to FormControls switched from undefined to string. This should not happen according to https://reactjs.org/docs/forms.html#controlled-components and they should always be strings.