feat(api/api-db-mongodb/editor): moved subscription queries to prisma
BREAKING CHANGE: SubscriptionList pagination rewritten. Now takes a cursor & take argument instead of first/last & before/after. Cursor is not base64 encoded anymore and skip argument is now the amount of items not pages.

BREAKING CHANGE: SubscriptionDeactivationReason is not an enum of integer anymore but is now an enum of string with the values `none` (used to be 0), `userSelfDeactivated` (used to be 1) and `invoiceNotPaid` (used to be 2)