refactor(editor): type changes, styling changes due to html5 doctype, serve only serving HTML by default now & fixed port/PORT mixmatch
Type changes:

Due to RSuite incorrectly typing rowData I created a new type and now used it everywhere.
This way it is correctly typed and latest TypeScript does not complain about it.

Styling changes:

Editor was lacking the HTML5 doctype before the move to Nx which meant that all HTML5 elements were lacking their functionality and base styling.
Due to the `display` property now being there and the effect it has on CSS I had to change a few things as height: 100% didn't work as before.

Serve changes:

I identified a memory leak with the @webpack-ssr-server which was also the cause for not properly updating the page after a change has happened.
As the SSR server is only for overriding environment variables and not needed for local development, I switched the default to be static files only.

Port changes:

I previously changed the docker ports from 8000 to 3000 as the editor server was running on 3000. I identified this as a case sensitivity issue and changed
port to PORT. This meant I could revert my changes from earlier.