feat(editor): migrated to nx (#789)
* chore: integrate nx into monorepo

1. Migrated node to v16 (https://wepublish.atlassian.net/browse/WPC-680)
2. Added required files for Nx and replaced our current eslint config with Nx's. tsconfig.base.old will be replaced at a later point when all projects have been migrated
3. Did some refactorings because of the changed eslint rules
4. Migrated webpack cli to 4 as we were using the 3.x CLI with version 4

* Update jest.preset.js

* chore: migrated apollo-server to v3

Required to migrate to GraphQL 15, which required GraphQLDate to be from graphql-scalars

* feat(editor): migrated to nx

 is now in

1. Codegen is now for the whole project
2. All tests now have a .test postfix in their filename
3. runServer is gone, is now just calling the main.js
4. Future will just use .env.local

* chore: remove target of docker images for now

* chore: lerna changes

* feat(editor): autofocus email input on login page (#787)

* ci: fix permissions missing for GA (#790)

* remove wrong return statement in Apollo error handler (#794)

* fix/WPC-951-matomo-multi-instances (#792)

* implement new wepublish-analytics package

* update wep-analytics-package

* refactor(editor): type changes, styling changes due to html5 doctype, serve only serving HTML by default now & fixed port/PORT mixmatch

Type changes:

Due to RSuite incorrectly typing rowData I created a new type and now used it everywhere.
This way it is correctly typed and latest TypeScript does not complain about it.

Styling changes:

Editor was lacking the HTML5 doctype before the move to Nx which meant that all HTML5 elements were lacking their functionality and base styling.
Due to the `display` property now being there and the effect it has on CSS I had to change a few things as height: 100% didn't work as before.

Serve changes:

I identified a memory leak with the @webpack-ssr-server which was also the cause for not properly updating the page after a change has happened.
As the SSR server is only for overriding environment variables and not needed for local development, I switched the default to be static files only.

Port changes:

I previously changed the docker ports from 8000 to 3000 as the editor server was running on 3000. I identified this as a case sensitivity issue and changed
port to PORT. This meant I could revert my changes from earlier.

Co-authored-by: Damian Janus <damian.janus@gmail.com>
Co-authored-by: michael-scheurer <70939054+michael-scheurer@users.noreply.github.com>