GRUP-12: Allow linking image blocks to a URL
This change allows adding the "linkUrl" to image blocks, which are then turned
into hyperlinks to the specified URL. This allows turning images into ads that
point to an external site.
The new linkUrl attribute is a bit of special case. All other content blocks
allow changing the block attributes with the corresponding EditPanel. The image
block is a bit different in that regard, and doesn't open the block attributes,
but those of the underlying image.
This PR modifies the EditPanel to also expose the new linkUrl attribute in
addition to the existing image attributes.