WPC-1419: Add page info to full text search
This adds the values required for pagination to the result of the phrase query.
In order to display a pagination component, at least the total number of
results is required. A first query is run without any pagination settings
applied. This produces a list of ids, and indirectly the total count of
matches. For loading the content, another query is run with the pagination
and sort orders applied.