SHARE-9: Allow media to inject custom Javascript code
With this change, media can enter custom Javascript code, for example for
displaying ads or embedding analytics solutions, into their editor. The
websites can embed the code by using one of the static script URLs, where
the code from the editor is automatically served as a plain Javascript file.
There is one setting for code that will be injected into the head, and one for
code for the body tag.

The scripts are served separately from the API, but from the same server for
two reasons:

1) Serving the code via GraphQL would make handling difficult and would require
parsing the response, extracting the code, and embedding it into the page.
Serving it directly as a "virtual" JS file makes embedding much easier.

2) This approach allows the scripts to be cached independently of the API
calls, similar to a static asset.

For serving the scripts, the Express instance that hosts the API is used,
and the scripts are loaded using an extra Express router.