SHARE-10: Make tests independent of current date
The snapshots of the SubscriptionEditView were broken because they depended on
a date picker that displayed the current date. When running the tests on a date
other than 2024-05-21, the rendered DOM was different than the snapshotted
version and the tests failed. This change mocks the date to a static date and
makes the tests work.