SHARE-9: Allow injecting complete scripts into the pages
This change allows users to not only inject script contents (or pure JS code)
into the pages, but to inject full scripts, including the wrapping <script>
tag. This way, attributes defined on the tag itself are preserved, such as the
data-id and the defer attributes of <script data-id="1234" defer></script>.

In order for the scripts to actually be run, the script must be recreated in
the frontend and all attributes must be copied. This is a bit ugly, but seems
to be the only way to preserve all original element data.