BA-33: Bajour Banner
This commit changes the appearance of the banner for Bajour. It allows all
actions to have a role. If the role is 'cancel', the action behaves the same as
the close button of the banner and minimizes it. If the role is 'other', the
action is displayed when the banner is maximized, but hidden if it is collapsed.
If the role is 'primary', the action is always displayed, no matter if collapsed
or expanded.

Additionally, the style is adapted to match Bajour's design.