BA-33: Change CTA behaviour (#1652)
This changes how the cta property of the banner works. When a cta is defined,
it is hidden when the banner is maximized, and is only shown when the banner is
minimized.

That means, if the banner is maximized, the title, the text, and all buttons
are shown.
If the banner is minimized, the CTA and the primary button are shown.