test(website): remove snapshots of components
Snapshots of components have become fairly useless now that we have a lot of them, a better way would be to use actual screenshots via storybook (chromatic).
After talking to Lukas him having the same opinion I got rid of them.

This PR is more to show intention, this will be merged at a different time as this would lead to a lot of merge conflicts right now.