feat(api): improve loading times of revisions by introducing archivedAt
Adding `archivedAt` to revisions allows us to always fetch the last 3 revisions because:

1. There can only be 1 pending revision at once
2. There can only be 1 draft revision not being archived
3. We sort by "publishedAt" descending

If we are loading 5 articles we used to have 16 queries. 1 for loading 5 articles and 3 per article for the revisions.
Now if we load 5 articles, we are doing 6 queries. 1 for loading 5 articles and 1 per article for the revisions.

The same applies to pages.

This also fixes the `draft` filter on articles & pages as we can now check if there's an un-archived draft version.

```description
Increase loading speed of articles & pages
Fixing the draft filter for articles & pages
```