feat(all): Restore full text search
Full text search requires us to use raw SQL queries in order for the tsvector
operators to work. Previously, to increase performance, the block content of
the articles and pages was stored in a separate column of the respective table.
This required every change to the blocks to update the search column as well.
With this commit, the search column is removed in favour of a full text index
on the blocks themselves. The full text search is extracted in the article-
and page service, and the phrase service calls them. This removes the burden
of remembering to update the search column, and centralizes the full text
search in the services, while keeping the general structure of the article
filters as they were.