feat(tsri): Add delay and html to banners (#1864)
This allows media to add a delay before displaying a banner. If the delay is
set to 0, the banner is shown immediately. Otherwise, it is delayed by the
specified number of seconds. The database has a default value of 0, so all
existing banners keep their behaviour.

Additionally, media can define a HTML content, which overrides the title and
text defined in the editor and allows them to show arbitrary markup.