fix(bajour): Fix the search slider search (#2024)
Using the articles query with both the title and body filter doesn't work, because
the backend then searches for articles having the search term in both the body and
the title, which almost never happens. Only using the body filter fixes that, since
it uses the full text search, which matches either title or body.