refactor(api): make all guids strings on a database level
As figured out by @dsheyp, when one tries to query an id that is not a guid/uuid, the database throws an error.
This is not desired in my opinion in applications like these as we map urls to ids, this keeps the ids being the same structure but with a more loose database type.