fix(website): Separate comment average and user rating
Since the average rating and the user rating are currently implemented in a
single rating component, it is impossible to rate the current average value,
since the component thinks the vote is already cast.
This changes the component to display two distinct rating components. One
displays the current average and is read only. The other appears as soon as the
user hovers over the ratings and allows them to cast a new rating. They are
visually distinct so it's easy to see if there already is a user rating or not.