feat: add external rewards to memberplans/subscriptions
Once a user has subscribed to a memberplan with an external reward, it will show in the user profile.
It requires the subscription to be active (startsAt in the past, paidUntil in the future or within grace period) before being able to see it.

Admins can always see the reward