fix(bajour): Fix search slider search
While the article count was returned correctly from the API, the actual article list was empty. This happened because the website included the `cursorId` variable in the query to only show articles after the currently selected one. Because the currently selected article was not in the result set, no articles were returned.
`cursorId` was still being included in the search query even though it was not present in the `searchArticles` function. This happens because Apollos `fetchMore` automatically merges the new variables with the original query's variables. Setting the variable explicitely to `undefined` skips it.